+++ src/main/java/warmup/one/WarmupOneTenthSet.java
@@ -30,4 +30,23 @@ public class WarmupOneTenthSet {
     public boolean lastDigit(int a, int b) {
         return b % 10 == a % 10;
     }
+
+    /**
+     * Given a string, return a new string where the last 3 chars are now in upper case. If the
+     * string has less than 3 chars, uppercase whatever is there. Note that str.toUpperCase()
+     * returns the uppercase version of a string.
+     *
+     * @param str
+     * @return
+     */
+    public String endUp(String str) {
+        if (str.length() > 3) {
+            int endIndex = str.length() - 3;
+            String endStr = str.substring(str.length()-3);
+
+            return str.substring(0, endIndex) + endStr.toUpperCase();
+        }
+
+        return str.toUpperCase();
+    }
 }